+++ src/__tests__/mergeProductWithItems.test.ts
@@ -1,6 +1,8 @@
 import { mergeProductWithItems } from '../mergeProductWithItems'
 import { itemsWithSimulationResponseMock } from './mock/itemsWithSimulationResponse'
+import { itemsWithSimulationResponse3PMock } from './mock/itemsWithSimulationResponse3P'
 import { vtexProductMock } from './mock/vtexProduct'
+import { vtexProduct3PMock } from './mock/vtexProduct3P'
 
 describe('mergeProductWithItems', () => {
   it('should be able to merge the results from the itemsWithSimulation query', () => {
@@ -27,5 +29,19 @@ describe('mergeProductWithItems', () => {
     expect(mergedSeller.commertialOffer.AvailableQuantity).toEqual(
       vtexProductMock.items[0].sellers[0].commertialOffer.AvailableQuantity
     )
+  })
+
+  it('should ignore simulation seller 1 if simulationBehavior only3P', () => {
+    const mergedProduct = mergeProductWithItems(vtexProduct3PMock, itemsWithSimulationResponse3PMock, 'only3P')
+
+    const seller1 = vtexProduct3PMock.items[0].sellers[0]
+
+    const [mergedItem] = mergedProduct.items
+    const mergedSellers = mergedItem.sellers
+    const seller1Merged = mergedSellers.find((seller) => seller.sellerId === '1')
+    const seller3PMerged = mergedSellers.find((seller) => seller.sellerId === 'seller3P')
+
+    expect(seller1Merged?.commertialOffer.spotPrice).toBe(375.77)
+    expect(seller3PMerged?.commertialOffer.spotPrice).toBe(223375.77)
   })
 })